代码错误修正，代码规范修正，注释添加

+++ frameworks/bridge/declarative_frontend/jsview/js_base_node.cpp
@@ -343,8 +343,8 @@ bool JSBaseNode::GetTouches(const JSCallbackInfo& info, TouchEvent& touchEvent)
             point.y = itemObj->GetPropertyValue<float>("y", 0.0f);
             point.screenX = itemObj->GetPropertyValue<float>("screenX", 0.0f);
             point.screenY = itemObj->GetPropertyValue<float>("screenY", 0.0f);
-            point.globalDisplayX = itemObj->GetPropertyValue<float>("globalDisplayX", 0.0f);
-            point.globalDisplayY = itemObj->GetPropertyValue<float>("globalDisplayY", 0.0f);
+            point.globalDisplayX = itemObj->GetPropertyValue<double>("globalDisplayX", 0.0);
+            point.globalDisplayY = itemObj->GetPropertyValue<double>("globalDisplayY", 0.0);
             point.originalId = itemObj->GetPropertyValue<int32_t>("id", 0);
             touchEvent.pointers.emplace_back(point);
         }
